Step 4: Change greet() to tell a random Chuck Joke
Update tests in accordance to the changes in behaviour of the World greet function. Note that now we've introduced and are using a dependency in our Java application, building a standalone Jar will no longer work, as the standalone Jar file will not be executable. This we'll address in next commit.